requirements.txt
alembic==1.6.2
aniso8601==9.0.1
beautifulsoup4==4.9.3
bidict==0.21.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==1.16.0
eventlet==0.30.2
firebase==3.0.1
Flask==1.1.2
Flask-JWT==0.3.2
Flask-Migrate==3.0.0
Flask-RESTful==0.3.8
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5.1
google==3.0.0
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-core==1.6.0
google-cloud-datastore==2.1.2
googleapis-common-protos==1.53.0
greenlet==1.0.0
grpcio==1.37.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
libcst==0.3.18
Mako==1.1.4
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.20.2
packaging==20.9
paho-mqtt==1.5.1
pandas==1.2.4
proto-plus==1.18.1
protobuf==3.15.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.13
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
Werkzeug==1.0.1
